+++ src/Domain/Queries/WordDetailsByWordQuery.cs
@@ -4,8 +4,8 @@ using Inshapardaz.Domain.Model;
 
 namespace Inshapardaz.Domain.Queries
 {
-    public class  WordDetailsByWordQuery : IQuery<IEnumerable<WordDetail>>
+    public class WordDetailsByWordQuery : IQuery<IEnumerable<WordDetail>>
     {
-        public int WordId { get; set; }
+        public long WordId { get; set; }
     }
 }
