+++ src/Domain/Queries/WordMeaningByWordQuery.cs
@@ -6,7 +6,7 @@ namespace Inshapardaz.Domain.Queries
 {
     public class WordMeaningByWordQuery : IQuery<IEnumerable<Meaning>>
     {
-        public int WordId { get; set; }
+        public long WordId { get; set; }
 
         public string Context { get; set; }
     }
